amara-ai/requirements.txt
crewai>=0.28.0
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
openai>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
azure-core>=1.32.0
azure-identity>=1.19.0
typing-extensions>=4.7.1
PyYAML>=6.0
langchain>=0.1.0
